Refactored top level and added upload_metadata
