+++ dpypelines/pipeline/validate_pipeline.py
@@ -3,10 +3,14 @@ import re
 from pathlib import Path
 from typing import Dict, List
 
+from dpytools.logging.logger import DpLogger
 from dpytools.validation.json.validation import validate_json_schema
 
+from dpypelines.pipeline.configuration import get_pipeline_config
 from dpypelines.pipeline.shared.pipelineconfig.matching import get_matching_pattern
 from dpypelines.pipeline.shared.utils import get_submitter_email
+
+logger = DpLogger("data-ingress-pipeline")
 
 
 def validate_pipeline_files(files_dir: Path, pipeline_config: dict) -> Dict:
@@ -119,3 +123,66 @@ def retrieve_and_validate_manifest(manifest_path: Path, required_keys: list) -> 
         return manifest_dict
     except Exception as e:
         raise ValueError(f"Failed to retrieve and validate manifest: {str(e)}")
+
+
+def retrieve_config_and_files(local_store):
+    """Retrieve configuration and files from the local directory."""
+    manifest_dict = retrieve_manifest(local_store)
+    source_id = get_source_id_from_manifest(manifest_dict)
+    pipeline_config = get_pipeline_config_for_source(source_id)
+    files_dir = Path(local_store.get_current_source_pathlike())
+
+    if not manifest_dict or not source_id or not pipeline_config or not files_dir:
+        err_msg = "Failed to retrieve configuration and files from the local directory."
+        logger.error(err_msg)
+        raise FileNotFoundError(err_msg)
+
+    logger.info(
+        "Configuration and files retrieved successfully",
+        data={
+            "manifest_dict": manifest_dict,
+            "source_id": source_id,
+            "pipeline_config": pipeline_config,
+            "files_dir": str(files_dir),
+        },
+    )
+
+    return manifest_dict, pipeline_config, files_dir
+
+
+def retrieve_manifest(local_store):
+    """Retrieve the manifest.json file from the local directory."""
+    manifest_dict = local_store.get_lone_matching_json_as_dict("manifest.json")
+
+    if not manifest_dict:
+        err_msg = "manifest.json not found in the local store."
+        logger.error(err_msg)
+        raise FileNotFoundError(err_msg)
+    return manifest_dict
+
+
+def get_source_id(manifest_dict: dict) -> str:
+    """
+    This function returns the `source_id` form the provided manifest_dict (which is the data in the manifest.json file).
+    """
+    return manifest_dict["source_id"]
+
+
+def get_source_id_from_manifest(manifest_dict):
+    """Extract the source_id from the manifest."""
+    source_id = get_source_id(manifest_dict)
+    if not source_id:
+        err_msg = f"source_id not found within manifest: {manifest_dict}."
+        logger.error(err_msg)
+        raise KeyError(err_msg)
+    return source_id
+
+
+def get_pipeline_config_for_source(source_id):
+    """Retrieve pipeline configuration using source_id."""
+    pipeline_config = get_pipeline_config(source_id)
+    if not pipeline_config:
+        err_msg = f"Pipeline configuration not found for source_id: {source_id}."
+        logger.error(err_msg)
+        raise ValueError(err_msg)
+    return pipeline_config
