+++ features/steps/dataset_ingress.py
@@ -3,15 +3,7 @@ import json
 import pandas as pd
 from behave import *
 from dictdiffer import diff
-from dpypelines.s3_folder_received import (
-    start,
-    setup_clients,
-    decompress_tar_file,
-    retrieve_config_and_files,
-    validate_pipeline,
-    upload_files,
-    send_submission_confirmation,
-)
+from dpypelines.s3_folder_received import start
 from unittest.mock import patch, MagicMock
 import os
